Restructure gadgets folder.
