+++ nova/src/circuits/nova/pcd/augmented.rs
@@ -27,9 +27,8 @@ use crate::{
         nimfs::{NIMFSProof, R1CSInstance, R1CSShape, RelaxedR1CSInstance},
         secondary::Circuit as SecondaryCircuit,
     },
-    gadgets::cyclefold::{
-        multifold, multifold_with_relaxed, primary, secondary, NonNativeAffineVar,
-    },
+    gadgets::cyclefold::{nova::{multifold, multifold_with_relaxed, primary}, secondary},
+    gadgets::nonnative::short_weierstrass::NonNativeAffineVar,
 };
 
 pub const SQUEEZE_NATIVE_ELEMENTS_NUM: usize = 1;
