+++ nova/src/circuits/supernova/augmented.rs
@@ -27,7 +27,8 @@ use crate::{
         nimfs::{NIMFSProof, R1CSInstance, RelaxedR1CSInstance},
         secondary::Circuit as SecondaryCircuit,
     },
-    gadgets::cyclefold::{multifold, primary, secondary, NonNativeAffineVar},
+    gadgets::cyclefold::{nova::{multifold, primary}, secondary},
+    gadgets::nonnative::short_weierstrass::NonNativeAffineVar,
 };
 
 pub const SQUEEZE_NATIVE_ELEMENTS_NUM: usize = 1;
